web端教师数据维护逻辑已更新，教师和学生绑定增加所属实验室认领逻辑

+++ springboot/src/main/java/ltd/syskaoqin/springboot/controller/admin/AdminTeacherManageController.java
@@ -3,9 +3,12 @@ package ltd.syskaoqin.springboot.controller.admin;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import ltd.syskaoqin.springboot.dao.entity.Teacher;
+import ltd.syskaoqin.springboot.dao.entity.UserAndLab;
 import ltd.syskaoqin.springboot.service.TeacherService;
+import ltd.syskaoqin.springboot.service.UserAndLabService;
 import ltd.syskaoqin.springboot.util.result.Result;
 import ltd.syskaoqin.springboot.util.result.ResultUtils;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
@@ -24,6 +27,8 @@ public class AdminTeacherManageController {
 
     @Resource
     private TeacherService teacherService;
+    @Resource
+    private UserAndLabService userAndLabService;
 
     @GetMapping(value = "/getTeacherList")
     @ResponseBody
@@ -41,8 +46,8 @@ public class AdminTeacherManageController {
 
     @GetMapping(value = "/checkTeacher")
     @ResponseBody
-    public Result checkTeacher(String num) {
-        Map<String,String> teacher = teacherService.checkTeacher(num);
+    public Result checkTeacher(String num,String labId) {
+        Map<String,String> teacher = teacherService.checkTeacher(num,labId);
         if(teacher == null){
             return ResultUtils.error(-1,"用户不存在或已添加");
         }else {
@@ -53,15 +58,29 @@ public class AdminTeacherManageController {
     @PostMapping(value = "/addTeacher")
     public Result addTeacher(@RequestBody Map<String, String> param){
         String num = param.get("num");
+        String labId = param.get("labId");
         Teacher teacher = teacherService.findTeacherInSys(num);
-        teacherService.insertTeacher(teacher);
+        UserAndLab userAndLab = new UserAndLab();
+        userAndLab.setStuNumber(num);
+        userAndLab.setLabId(labId);
+        try {
+            teacherService.insertTeacher(teacher);
+        }catch (DuplicateKeyException e){
+            e.printStackTrace();
+        }
+        userAndLabService.insertUserLabTea(userAndLab);
         return ResultUtils.success();
     }
 
     @PostMapping(value = "/deleteTeacher")
     public Result deleteTeacher(@RequestBody Map<String, String> param){
         String num = param.get("num");
-        teacherService.deleteTeacher(num);
+        String labId = param.get("labId");
+        userAndLabService.deleteByTeaNumber(num,labId);
+        List<UserAndLab> userAndLabList = userAndLabService.findByTeaNumber(num);
+        if(userAndLabList == null){
+            teacherService.deleteTeacher(num);
+        }
         return ResultUtils.success();
     }
 }
